Fix of a crash on startup when creating physical printer profiles from 2.2 configs (5513, 5074)
The issue is that opening the wxMessageDialog in GUI_App::check_printer_presets  fires wxEVT_IDLE event (on Linux at least).
At that point it is already bound to our lambda which will in turn call post_init prematurely.
The solution is to move the Bind at the very end of the initialization. The post_init that it calls assumes that init has been finished anyway,
registering sooner should not be needed. Is there some reason to register the event earlier when the post_init cannot be called yet?